+++ bin/classes/product.dart
