+++ bin/classes/stock.dart
@@ -1,4 +1,4 @@
-import 'Product.dart';
+import 'product.dart';
 
 class Stock {
   List<Product>? products;
