+++ bin/projeto_inicial.dart
@@ -1,7 +1,7 @@
 import 'classes/index.dart';
 import 'dart:io';
 
-import 'error/ErrorMessage.dart';
+import 'error/error_message.dart';
 
 void main() {
   // Mercado
